refactor: Remove redundant tests from .NET Standard project
