--- sdk/test/NetStandard/IntegrationTests/IntegrationTests/AWSSupport.cs
@@ -1,151 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-using Amazon.AWSSupport;
-using Amazon.AWSSupport.Model;
-using Amazon;
-using System.IO;
-using System.Text;
-using Amazon.DNXCore.IntegrationTests;
-using Xunit;
-
-namespace Amazon.DNXCore.IntegrationTests
-{
-    
-    public class AWSSupportTests : TestBase<AmazonAWSSupportClient>
-    {
-        private static String
-            SUBJECT = ".NET SDK Test Case " + DateTime.UtcNow.Ticks,
-            CATEGORY_CODE = "apis",
-            SERVICE_CODE = "amazon-dynamodb",
-            COMMUNICATION_BODY = "This is a test case generated by the .NET SDK integration test suite",
-            LANGUAGE = "ja",
-            SEVERITY_CODE = "low",
-            ATTACHMENT_CONTENTS = "This is test data";
-
-        protected override RegionEndpoint AlternateEndpoint
-        {
-            get
-            {
-                return RegionEndpoint.USEast1;
-            }
-        }
-        
-        //  Test are disabled because not all acounts are subscribed to AWS Support
-        //[Fact]
-        public async Task TestCaseOperations()
-        {
-            string caseId = null;
-
-            try
-            {
-                caseId = Client.CreateCaseAsync(new CreateCaseRequest
-                {
-                    Subject = SUBJECT,
-                    CategoryCode = CATEGORY_CODE,
-                    ServiceCode = SERVICE_CODE,
-                    Language = LANGUAGE,
-                    SeverityCode = SEVERITY_CODE,
-                    CommunicationBody = COMMUNICATION_BODY
-                }).Result.CaseId;
-
-                Assert.NotNull(caseId);
-
-                var cases = (await Client.DescribeCasesAsync(new DescribeCasesRequest { Language = LANGUAGE })).Cases;
-                Assert.True(cases.Count > 0);
-
-                cases = (await Client.DescribeCasesAsync(new DescribeCasesRequest { Language = LANGUAGE, CaseIdList = new List<string> { caseId } })).Cases;
-                Assert.Single(cases);
-
-                Assert.Equal(caseId, cases[0].CaseId);
-                Assert.Equal(CATEGORY_CODE, cases[0].CategoryCode);
-                Assert.Equal(LANGUAGE, cases[0].Language);
-                Assert.Equal(SERVICE_CODE, cases[0].ServiceCode);
-                Assert.Equal(SEVERITY_CODE, cases[0].SeverityCode);
-                Assert.True(cases[0].RecentCommunications.Communications.Count > 0);
-
-                var attachmentData = new MemoryStream(Encoding.UTF8.GetBytes(ATTACHMENT_CONTENTS));
-                var filename = "file1.txt";
-                var attachmentSetId = (await Client.AddAttachmentsToSetAsync(new AddAttachmentsToSetRequest
-                {
-                    Attachments = new List<Attachment>
-                    {
-                        new Attachment
-                        {
-                            FileName = filename,
-                            Data = attachmentData
-                        }
-                    }
-                })).AttachmentSetId;
-
-                var result = await Client.AddCommunicationToCaseAsync(new AddCommunicationToCaseRequest
-                {
-                    CaseId = caseId,
-                    CcEmailAddresses = new List<string> { "aws-dr-tools-test@amazon.com" },
-                    CommunicationBody = COMMUNICATION_BODY,
-                    AttachmentSetId = attachmentSetId
-                });
-
-                Assert.NotNull(result);
-
-                var comms = (await Client.DescribeCommunicationsAsync(new DescribeCommunicationsRequest { CaseId = caseId })).Communications;
-                Assert.True(comms.Count > 0);
-                Assert.Equal(caseId, comms[0].CaseId);
-                Assert.Equal(COMMUNICATION_BODY.Trim(), comms[0].Body.Trim());
-                Assert.NotNull(comms[0].SubmittedBy);
-                Assert.NotNull(comms[0].TimeCreated);
-
-                string attachmentId = null;
-                attachmentId = GetAttachmentId(comms, attachmentId);
-                Assert.NotNull(attachmentId);
-
-                await VerifyAttachment(attachmentData, filename, attachmentId);
-
-                cases = (await Client.DescribeCasesAsync(new DescribeCasesRequest { Language = LANGUAGE, CaseIdList = new List<string> { caseId }, IncludeCommunications = true })).Cases;
-                Assert.Single(cases);
-                var communications = cases[0].RecentCommunications;
-                attachmentId = GetAttachmentId(communications.Communications, attachmentId);
-                await VerifyAttachment(attachmentData, filename, attachmentId);
-            }
-            finally
-            {
-                if (caseId != null)
-                {
-                    await Client.ResolveCaseAsync(new ResolveCaseRequest { CaseId = caseId });
-                }
-            }
-        }
-
-        private async Task VerifyAttachment(MemoryStream attachmentData, string filename, string attachmentId)
-        {
-            var attachment = (await Client.DescribeAttachmentAsync(new DescribeAttachmentRequest
-            {
-                AttachmentId = attachmentId
-            })).Attachment;
-            Assert.NotNull(attachment);
-            Assert.Equal(
-                Encoding.UTF8.GetString(attachmentData.ToArray()),
-                Encoding.UTF8.GetString(attachment.Data.ToArray()));
-            Assert.Equal(filename, attachment.FileName);
-        }
-
-        private static string GetAttachmentId(List<Communication> comms, string attachmentId)
-        {
-            foreach (var comm in comms)
-            {
-                var attachmentSet = comm.AttachmentSet;
-                if (attachmentSet != null && attachmentSet.Count > 0)
-                {
-                    foreach (var att in attachmentSet)
-                    {
-                        if (!string.IsNullOrEmpty(att.AttachmentId))
-                            attachmentId = att.AttachmentId;
-                    }
-                }
-            }
-            return attachmentId;
-        }
-    }
-}
