+++ sdk/test/NetStandard/IntegrationTests/IntegrationTests/ProtocolTests.cs
@@ -1,18 +1,9 @@
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.Model;
-using Amazon.ElasticTranscoder;
-using Amazon.ElasticTranscoder.Model;
 using Amazon.Runtime;
 using Amazon.S3;
 using Amazon.S3.Model;
-using Amazon.SimpleWorkflow;
-using Amazon.SimpleWorkflow.Model;
-using Amazon.DNXCore.IntegrationTests;
 using Xunit;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Amazon.DNXCore.IntegrationTests
@@ -23,53 +14,10 @@ namespace Amazon.DNXCore.IntegrationTests
     
     public class ProtocolTests
     {
-        //[Fact]
-        public void TestJson()
-        {
-            UtilityMethods.RunAsSync(TestJson_Async);
-        }
-
-        private async Task TestJson_Async()
-        {
-            using (var client = UtilityMethods.CreateClient<AmazonSimpleWorkflowClient>())
-            {
-                var domainName = "dotnet-test-domain-" + DateTime.UtcNow.ToFileTime();
-
-                await client.RegisterDomainAsync(new RegisterDomainRequest
-                {
-                    Name = domainName,
-                    Description = "Test domain",
-                    WorkflowExecutionRetentionPeriodInDays = "5"
-                });
-
-                var domains = (await client.ListDomainsAsync(new ListDomainsRequest
-                {
-                    RegistrationStatus = RegistrationStatus.REGISTERED
-                })).DomainInfos;
-
-                Assert.NotNull(domains);
-                Assert.NotNull(domains.Infos);
-                Assert.NotEmpty(domains.Infos);
-
-                await client.DeprecateDomainAsync(new DeprecateDomainRequest
-                {
-                    Name = domainName
-                });
-
-                var ure = await AssertExtensions.ExpectExceptionAsync<UnknownResourceException>(client.DeprecateDomainAsync(new DeprecateDomainRequest
-                {
-                    Name = "really-fake-domain-that-should-not-exist" + DateTime.UtcNow.ToFileTime()
-                }));
-                Assert.NotNull(ure);
-                Assert.NotNull(ure.Message);
-                Assert.NotNull(ure.ErrorCode);
-                Assert.Equal(ErrorType.Unknown, ure.ErrorType);
-            }
-        }
-
         [Fact]
         public async Task TestRestXml()
         {
+            // TODO: Validate this test is still necessary after the S3 codegen migration is complete.
             using (var client = UtilityMethods.CreateClient<AmazonS3Client>())
             {
                 var bucketName = await UtilityMethods.CreateBucketAsync(client, "TestRestXml_Async");
@@ -87,9 +35,9 @@ namespace Amazon.DNXCore.IntegrationTests
                     {
                         BucketName = fakeBucketName
                     }));
+
                     Assert.NotNull(as3e);
                     Assert.NotNull(as3e.Message);
-                    //Assert.NotNull(aete.ErrorCode);
                     Assert.Equal(ErrorType.Sender, as3e.ErrorType);
                 }
                 finally
